pyproject.toml
[tool.poetry]
name = "intervista-assistant"
version = "0.1.0"
description = "AI-powered interview assistant using Google Gemini"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
flask = "^2.0.1"
flask-cors = "^4.0.0"
flask-socketio = "^5.3.0"
python-socketio = "^5.7.2"
python-dotenv = "^1.0.0"
google-generativeai = "^0.3.0"
numpy = "^1.24.0"
pillow = "^10.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start-api = "intervista_assistant.api_launcher:main"
